docs: add tests and readme
