+++ package.json
@@ -26,17 +26,23 @@
   "scripts": {
     "build": "rollup -c",
     "watch": "rollup -cw",
-    "prepublishOnly": "yarn build"
+    "prepublishOnly": "yarn build",
+    "test": "jest"
   },
   "dependencies": {},
   "devDependencies": {
     "@rollup/plugin-node-resolve": "^13.0.5",
+    "@types/jest": "^27.4.1",
+    "@types/node": "^14.11.8",
     "@types/unist": "^2.0.6",
     "concurrently": "^6.2.1",
+    "jest": "^27.5.1",
     "rollup": "^2.57.0",
     "rollup-plugin-delete": "^2.0.0",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-typescript2": "^0.30.0",
+    "ts-jest": "^27.1.4",
+    "ts-node": "^10.7.0",
     "typescript": "^4.4.3"
   },
   "peerDependencies": {
